Comment tidying and one correctness fix across all source files
- Convert C++ // comments to /* */ throughout (scsi_parse.c, Close.c,
  scsi_log_sense.c, cmd_td_getnumtracks.c)
- Fix IUtility->Strlen() (does not exist) -> strlen() in scsi_log_sense.c
- Fix TD_GETDRIVETYPE in BeginIO.c returning 0 instead of DRIVE_NEWSTYLE (0x44)
- Split BeginIO.c inline cases with accurate per-case value comments
- Merge duplicate EVENT_IDX/INDIRECT_DESC block comments in virtio_init.c
- Move unreachable NOTE comment above return 0 in VirtIOSCSI_Submit()
- Merge orphaned one-liner into UnitTask_Dispatch() block comment header
- Tighten pci_discovery.c comments; remove now-unused 'found' variable
- Reorder SCSI opcode cases in scsi_parse.c to numeric order

+++ src/exec_cmds/cmd_td_getnumtracks.c
@@ -4,10 +4,10 @@
 void Handle_TD_GetNumTracks(struct VirtIOSCSIBase *libBase, struct IOStdReq *req)
 {
     struct VirtIOUSCSIDevUnit *unit = (struct VirtIOUSCSIDevUnit *)req->io_Unit;
-    uint32 cylinders = 32768; // Safe default for large hard disks
+    uint32 cylinders = 32768; /* safe default for large disks */
 
     if (unit && unit->geometry_valid) {
-        cylinders = unit->total_blocks / (4 * 16); // heads * sectors_per_track
+        cylinders = (uint32)(unit->total_blocks / (4 * 16)); /* heads=4, spt=16 */
     }
 
     DPRINTF(libBase->IExec, "[virtioscsi] TD_GETNUMTRACKS: Returning %lu for T%lu L%lu (Valid: %d)\n", cylinders,
